requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
httpx==0.27.2
apscheduler==3.10.4
pydantic==2.8.2
pydantic-settings==2.4.0
pytest==8.3.2
python-dotenv==1.0.1
